requirements.txt
py2app
PyAutoGUI
PyQt5
python-socketio
eventlet